+++ app/src/main/java/woowacourse/shopping/ui/util/FormattedPrice.kt
@@ -2,7 +2,7 @@ package woowacourse.shopping.ui.util
 
 import java.text.DecimalFormat
 
-fun formattedPrice(value: Int): String {
+fun formattedPrice(value: Long): String {
     val decimal = DecimalFormat("#,###원")
     return decimal.format(value)
 }
